NEW ROUTING

+++ GraderApi/Controllers/CourseUsersController.cs
@@ -3,7 +3,6 @@
     using Grader.JsonSerializer;
     using GraderDataAccessLayer.Interfaces;
     using GraderDataAccessLayer.Models;
-    using GraderDataAccessLayer.Repositories;
     using System;
     using System.Collections.Generic;
     using System.Net;
@@ -11,7 +10,6 @@
     using System.Threading.Tasks;
     using System.Web.Http;
     using System.Web.Http.Description;
-
 
     public class CourseUsersController : ApiController
     {
@@ -21,9 +19,10 @@
             _courseUserRepository = courseUserRepository;
         }
 
-        // GET: api/CourseUsers
+        // GET: api/CourseUsers/All
         [HttpGet]
         [ResponseType(typeof(IEnumerable<CourseUserModel>))]
+        [PermissionsAuthorize(SuperUserPermissions.CanSeeAllCourseUsers)]
         public async Task<HttpResponseMessage> All()
         {
             try
@@ -37,17 +36,15 @@
             }
         }
 
-        // GET: api/CourseUsers/5
+        // GET: api/Courses/{courseId}/CourseUsers
         [HttpGet]
-        [ResponseType(typeof(CourseUserModel))]
-        public async Task<HttpResponseMessage> Get(int courseUserId)
+        [ResponseType(typeof(IEnumerable<CourseUserModel>))]
+        public async Task<HttpResponseMessage> All(int courseId)
         {
             try
             {
-                var courseUser = await _courseUserRepository.Get(courseUserId);
-
-                return courseUser != null ? Request.CreateResponse(HttpStatusCode.Accepted, courseUser.ToJson())
-                    : Request.CreateResponse(HttpStatusCode.NotFound);
+                var result = await _courseUserRepository.GetAllByCourseId(courseId);
+                return Request.CreateResponse(HttpStatusCode.OK, result.ToJson());
             }
             catch (Exception e)
             {
@@ -55,15 +52,44 @@
             }
         }
 
-        // POST: api/CourseUsers
+        // GET: api/Courses/{courseId}/CourseUsers/{courseUserId}
+        [HttpGet]
+        [ResponseType(typeof(CourseUserModel))]
+        [PermissionsAuthorize(CourseOwnerPermissions.CanSeeEnrollment)]
+        public async Task<HttpResponseMessage> Get(int courseId, int courseUserId)
+        {
+            try
+            {
+                var courseUser = await _courseUserRepository.Get(courseUserId);
+                if (courseUser == null)
+                {
+                    return Request.CreateResponse(HttpStatusCode.NotFound);
+                }
+
+
+                return courseId == courseUser.CourseId
+                    ? Request.CreateResponse(HttpStatusCode.Accepted, courseUser.ToJson())
+                    : Request.CreateResponse(HttpStatusCode.BadRequest, Messages.InvalidCourse);
+            }
+            catch (Exception e)
+            {
+                return Request.CreateErrorResponse(HttpStatusCode.InternalServerError, e);
+            }
+        }
+
+        // POST: api/Courses/{courseId}/CourseUsers
         [HttpPost]
         [ResponseType(typeof(CourseUserModel))]
         [PermissionsAuthorize(CourseOwnerPermissions.CanAddEnrollment)]
-        public async Task<HttpResponseMessage> Add([FromBody] CourseUserModel courseUser)
+        public async Task<HttpResponseMessage> Add(int courseId, [FromBody] CourseUserModel courseUser)
         {
             if (!ModelState.IsValid)
             {
                 return Request.CreateResponse(HttpStatusCode.BadRequest, ModelState);
+            }
+            if (courseId != courseUser.CourseId)
+            {
+                return Request.CreateResponse(HttpStatusCode.BadRequest, Messages.InvalidCourse);
             }
 
             try
@@ -79,20 +105,23 @@
             }
         }
 
-        // PUT: api/CourseUsers/5
+        // PUT: api/Courses/{courseId}/CourseUsers/{courseUserId}
         [HttpPut]
         [ResponseType(typeof(CourseUserModel))]
         [PermissionsAuthorize(CourseOwnerPermissions.CanUpdateEnrollment)]
-        public async Task<HttpResponseMessage> Update(int courseUserId, [FromBody] CourseUserModel courseUser)
+        public async Task<HttpResponseMessage> Update(int courseId, int courseUserId, [FromBody] CourseUserModel courseUser)
         {
             if (!ModelState.IsValid)
             {
                 return Request.CreateResponse(HttpStatusCode.BadRequest, ModelState);
             }
-
             if (courseUserId != courseUser.Id)
             {
-                return Request.CreateResponse(HttpStatusCode.BadRequest);
+                return Request.CreateResponse(HttpStatusCode.BadRequest, Messages.InvalidCourse);
+            }
+            if (courseId != courseUser.CourseId)
+            {
+                return Request.CreateResponse(HttpStatusCode.BadRequest, Messages.InvalidCourse);
             }
 
             try
@@ -108,14 +137,24 @@
             }
         }
 
-        // DELETE: api/CourseUsers/5
+        // DELETE: api/Courses/{courseId}/CourseUsers/{courseUserId}
         [HttpDelete]
         [ResponseType(typeof(void))]
         [PermissionsAuthorize(CourseOwnerPermissions.CanDeleteEnrollment)]
-        public async Task<HttpResponseMessage> Delete(int courseUserId)
+        public async Task<HttpResponseMessage> Delete(int courseId, int courseUserId)
         {
             try
             {
+                var existingCourseUser = await _courseUserRepository.Get(courseUserId);
+                if (existingCourseUser == null)
+                {
+                    return Request.CreateResponse(HttpStatusCode.NotFound);
+                }
+                if (existingCourseUser.CourseId != courseId)
+                {
+                    return Request.CreateResponse(HttpStatusCode.BadRequest, Messages.InvalidCourse);
+                }
+
                 var result = await _courseUserRepository.Delete(courseUserId);
                 return Request.CreateResponse(!result ? HttpStatusCode.InternalServerError : HttpStatusCode.OK);
             }
