+++ GraderApi/Controllers/GradeComponentsController.cs
@@ -3,7 +3,6 @@
     using Grader.JsonSerializer;
     using GraderDataAccessLayer.Interfaces;
     using GraderDataAccessLayer.Models;
-    using GraderDataAccessLayer.Repositories;
     using System;
     using System.Collections.Generic;
     using System.Net;
@@ -11,7 +10,6 @@
     using System.Threading.Tasks;
     using System.Web.Http;
     using System.Web.Http.Description;
-
 
     public class GradeComponentsController : ApiController
     {
@@ -21,9 +19,10 @@
             _gradeComponentRepository = gradeComponentRepository;
         }
 
-        // GET: api/GradeComponents
+        // GET: api/GradeComponents/All
         [HttpGet]
-        [ResponseType(typeof(IEnumerable<GradeComponentModel>))]
+        [ResponseType(typeof (IEnumerable<GradeComponentModel>))]
+        [PermissionsAuthorize(SuperUserPermissions.CanSeeAllGradedParts)]
         public async Task<HttpResponseMessage> All()
         {
             try
@@ -37,19 +36,40 @@
             }
         }
 
-        // GET: api/GradeComponents/5
+        // GET: api/Courses/{courseId}/GradeComponents/All
+        [HttpGet]
+        [ResponseType(typeof(IEnumerable<GradeComponentModel>))]
+        [PermissionsAuthorize(CoursePermissions.CanSeeGradedParts)]
+        public async Task<HttpResponseMessage> All(int courseId)
+        {
+            try
+            {
+                var result = await _gradeComponentRepository.GetAllByCourse(courseId);
+                return Request.CreateResponse(HttpStatusCode.OK, result.ToJson());
+            }
+            catch (Exception e)
+            {
+                return Request.CreateErrorResponse(HttpStatusCode.InternalServerError, e);
+            }
+        }
+
+        // GET: api/Courses/{courseId}/GradeComponents/{gradeComponentId}
         [HttpGet]
         [ResponseType(typeof(GradeComponentModel))]
-        [PermissionsAuthorize(CoursePermissions.CanSeeGrades)]
-        public async Task<HttpResponseMessage> Get(int gradeComponentId)
+        [PermissionsAuthorize(CoursePermissions.CanSeeGradedParts)]
+        public async Task<HttpResponseMessage> Get(int courseId, int gradeComponentId)
         {
             try
             {
                 var gradeComponent = await _gradeComponentRepository.Get(gradeComponentId);
+                if (gradeComponent == null)
+                {
+                    return Request.CreateResponse(HttpStatusCode.NotFound);   
+                }
 
-                return gradeComponent != null
+                return gradeComponent.CourseId == courseId
                     ? Request.CreateResponse(HttpStatusCode.OK, gradeComponent.ToJson())
-                    : Request.CreateResponse(HttpStatusCode.NotFound);
+                    : Request.CreateResponse(HttpStatusCode.BadRequest, Messages.InvalidCourse);
             }
             catch (Exception e)
             {
@@ -57,55 +77,19 @@
             }
         }
 
-        // GET: /api/GradeComponents/GetCourse/3
-        [HttpGet]
-        [ResponseType(typeof (CourseModel))]
-        [PermissionsAuthorize(CoursePermissions.CanSeeGrades)]
-        public async Task<HttpResponseMessage> GetCourse(int gradeComponentId)
-        {
-            try
-            {
-                var gradeComponent = await _gradeComponentRepository.Get(gradeComponentId);
-
-                return gradeComponent != null 
-                    ? Request.CreateResponse(HttpStatusCode.OK, gradeComponent.Course.ToJson()) 
-                    : Request.CreateResponse(HttpStatusCode.NotFound);
-            }
-            catch (Exception e)
-            {
-                return Request.CreateErrorResponse(HttpStatusCode.InternalServerError, e);
-            }
-        }
-
-        // GET: /api/GradeComponents/GetForCourse/3
-        [HttpGet]
-        [ResponseType(typeof (IEnumerable<GradeComponentModel>))]
-        [PermissionsAuthorize(CoursePermissions.CanSeeGrades)]
-        public async Task<HttpResponseMessage> GetForCourse(int courseId)
-        {
-            try
-            {
-                var gradeComponents = await _gradeComponentRepository.GetAllByCourse(courseId);
-
-                return gradeComponents != null
-                    ? Request.CreateResponse(HttpStatusCode.OK, gradeComponents.ToJson())
-                    : Request.CreateResponse(HttpStatusCode.NotFound);
-            }
-            catch (Exception e)
-            {
-                return Request.CreateErrorResponse(HttpStatusCode.InternalServerError, e);
-            }
-        }
-
-        // POST: api/GradeComponents
+        // POST: api/Courses/{courseId}/GradeComponents
         [HttpPost]
         [ResponseType(typeof(GradeComponentModel))]
         [PermissionsAuthorize(CourseOwnerPermissions.CanCreateGradedPart)]
-        public async Task<HttpResponseMessage> Add([FromBody] GradeComponentModel gradeComponent)
+        public async Task<HttpResponseMessage> Add(int courseId, [FromBody] GradeComponentModel gradeComponent)
         {
             if (!ModelState.IsValid) 
             {
                 return Request.CreateResponse(HttpStatusCode.BadRequest, ModelState);
+            }
+            if (courseId != gradeComponent.CourseId)
+            {
+                return Request.CreateResponse(HttpStatusCode.BadRequest, Messages.InvalidCourse);
             }
 
             try
@@ -121,20 +105,23 @@
             }
         }
 
-        // PUT: api/GradeComponents/5
+        // PUT: api/Courses/{courseId}/GradeComponents/{gradeComponentId}
         [HttpPut]
         [ResponseType(typeof(GradeComponentModel))]
         [PermissionsAuthorize(CourseOwnerPermissions.CanUpdateGradedPart)]
-        public async Task<HttpResponseMessage> Update(int gradeComponentId, [FromBody] GradeComponentModel gradeComponent)
+        public async Task<HttpResponseMessage> Update(int courseId, int gradeComponentId, [FromBody] GradeComponentModel gradeComponent)
         {
             if (!ModelState.IsValid) 
             {
                 return Request.CreateResponse(HttpStatusCode.BadRequest, ModelState);
             }
-
             if (gradeComponentId != gradeComponent.Id) 
             {
-                return Request.CreateResponse(HttpStatusCode.BadRequest);
+                return Request.CreateResponse(HttpStatusCode.BadRequest, Messages.InvalidGradeComponentId);
+            }
+            if (courseId != gradeComponent.CourseId)
+            {
+                return Request.CreateResponse(HttpStatusCode.BadRequest, Messages.InvalidCourse);
             }
 
             try
@@ -151,14 +138,24 @@
             }
         }
 
-        // DELETE: api/GradeComponents/5
+        // DELETE: api/Courses/{courseId}/GradeComponents/{gradeComponentId}
         [HttpDelete]
         [ResponseType(typeof(void))]
         [PermissionsAuthorize(CourseOwnerPermissions.CanDeleteGradedPart)]
-        public async Task<HttpResponseMessage> Delete(int gradeComponentId)
+        public async Task<HttpResponseMessage> Delete(int courseId, int gradeComponentId)
         {
             try
             {
+                var existingGradeComponent = await _gradeComponentRepository.Get(gradeComponentId);
+                if (existingGradeComponent == null)
+                {
+                    return Request.CreateResponse(HttpStatusCode.NotFound);
+                }
+                if (existingGradeComponent.CourseId != courseId)
+                {
+                    return Request.CreateResponse(HttpStatusCode.BadRequest, Messages.InvalidCourse);
+                }
+
                 var result = await _gradeComponentRepository.Delete(gradeComponentId);
                 return Request.CreateResponse(result ? HttpStatusCode.OK : HttpStatusCode.InternalServerError);
             }
